internal/handlers/auth_handler.go
package handler

import (
	"github.com/gin-gonic/gin"
	"net/http"
	"taxi-service/utils"
)

func LoginHandler(c *gin.Context) {
	var credentials struct {
		Username string `json:"username" binding:"required"`
		Password string `json:"password" binding:"required"`
	}

	if err := c.ShouldBindJSON(&credentials); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid input"})
		return
	}

	// Replace this with proper username and password validation
	if credentials.Username == "john" && credentials.Password == "setup123" {
		token, err := utils.GenerateToken(credentials.Username)
		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{"error": "Could not generate token"})
			return
		}
		c.JSON(http.StatusOK, gin.H{"token": token})
		return
	}

	c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid username or password"})
}
